WE-24 언어 선택하는 타이틀부분은 별도의 컴포넌트로 분리 WE-104
